Moved MatchStateMachine into it's own file
