+++ engine/Assets/Scripts/Modes/MatchStateMachine.cs
@@ -1,0 +1,58 @@
+﻿using Synthesis.UI.Dynamic;
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using TMPro;
+
+public class MatchStateMachine {
+    public static int currentFieldIndex = -1;
+    public static int currentRobotIndex = -1;
+
+    public const string PREVIOUS_SPAWN_LOCATION = "Previous Spawn Location";
+    public const string PREVIOUS_SPAWN_ROTATION = "Previous Spawn Rotation";
+    
+    private MatchState _currentState;
+
+    public MatchState CurrentState
+    {
+        get
+        {
+            return _currentState;
+        }
+        set 
+        {
+            Debug.Log($"State set to {value}");
+            if (value == _currentState)
+            {
+                Debug.LogError($"New state is the same as the current state ({value})");
+                return; 
+            }
+            
+            _currentState = value;
+            switch (value)
+            {
+                case MatchState.MatchConfig:
+                    DynamicUIManager.CreateModal<MatchModeModal>();
+                    ((MatchModeModal)DynamicUIManager.ActiveModal).OnAccepted += () => CurrentState = MatchState.RobotPositioning;
+                    return; 
+                case MatchState.RobotPositioning:
+                    DynamicUIManager.CreatePanel<StartMatchModePanel>();
+                    return;
+                case MatchState.Auto: 
+                    return;
+                default:
+                    Debug.LogError($"No behavior for current state ({value})");
+                    return;
+            }
+        }
+    }
+    
+
+    public enum MatchState
+    {
+        None,
+        MatchConfig,
+        RobotPositioning,
+        Auto
+    }
+}
